requirements.txt
dj_rest_auth==6.0.0
Django==5.1.1
django_allauth==65.0.2
django_environ==0.11.2
django_parler_rest==2.2
djangorestframework_simplejwt==5.3.1
drf_yasg==1.21.7
Faker==30.3.0
PyJWT==2.10.1
python-dotenv==1.0.1
Requests==2.32.3
xhtml2pdf==0.2.16
django-jazzmin==3.0.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.3
django-jazzmin==3.0.1
django-js-asset==2.2.0
django-modeltranslation==0.19.9
django-mptt==0.16.0
django-parler==2.3
django-parler-rest==2.2
django-rest-swagger==2.2.0
django-rosetta==0.10.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
pillow==10.4.0
decorator==5.1.1
coreapi==2.3.3
coreschema==0.0.4
arabic-reshaper==3.0.0
python-decouple